requirements.txt
beautifulsoup4==4.5.1
coreapi==2.1.1
cssmin==0.2.0
defusedxml==0.4.1
dj-database-url==0.4.1
Django==1.10.2
django-allauth==0.30.0
django-app-namespace-template-loader==0.4.1
django-blog-zinnia==0.18.1
django-braces==1.10.0
django-ckeditor==5.1.1
django-contact-form==1.2
django-contrib-comments==1.7.3
django-cors-headers==2.0.1
django-debug-toolbar==1.6
django-feedparser==0.2.1
django-filetransfers==0.1.0
django-filter==1.0.1
django-grappelli==2.9.1
django-guardian==1.4.6
django-jet==1.0.4
django-mailer==1.2.2
django-material==0.11.0
django-model-utils==2.6
django-mptt==0.8.7
django-multiselectfield==0.1.5
django-notifications-hq==1.2
django-postman==3.5.1
django-rest-auth==0.9.0
django-star-ratings==0.5.3
django-tagging==0.4.5
django-taggit==0.21.3
django-threadedcomments==1.0.1
django-tinymce==2.4.0
django-userena==2.0.1
django-xmlrpc==0.1.7
djangorestframework==3.5.3
docx==0.2.4
drf-url-filters==0.2.0
easy-thumbnails==2.3
feedparser==5.2.1
gunicorn==19.6.0
html2text==2014.12.29
itypes==1.1.0
Jinja2==2.8
jsonfield==1.0.3
lockfile==0.12.2
lxml==3.7.3
Markdown==2.6.7
MarkupSafe==0.23
mots-vides==2015.5.11
notifications==0.3.2
oauthlib==2.0.1
Pillow==3.4.1
pyparsing==2.1.10
python-docx==0.8.6
python3-openid==3.0.10
pytz==2016.10
PyYAML==3.12
regex==2016.11.21
render==1.0.0
requests==2.12.4
requests-oauthlib==0.7.0
six==1.10.0
sorl-thumbnail==12.3
sqlparse==0.2.2
tweepy==3.5.0
uritemplate==3.0.0
virtualenv==15.0.3
virtualenvwrapper-win==1.2.1
voluptuous==0.9.3
whitenoise==3.2.2
zinnia-theme-bootstrap==0.5.1
zinnia-twitter==1.1
zinnia-wysiwyg-ckeditor==1.3
zinnia-wysiwyg-tinymce==1.3
